EE/VU: fix Tekken Tag hang from 64-bit cycle counter (cast regression)
Followup fix to the 64-bit cycle counter change. Two cycle-delta casts
were widened to (s64) when they must stay truncating, which broke EE event
scheduling and hung Tekken Tag Tournament at the Namco logo.

The EE event scheduler computes every cycle delta as (int)(a - b) and
relies on the low-32-bit truncation: because event deltas are always small,
the low 32 bits hold the correct signed delta even when the start cycle is
slightly ahead of the current cycle (the wraparound trick). All scheduler
comparisons in R5900.cpp keep (int) for this reason. The 64-bit conversion
should keep (int)/(s32) on these small-delta and M-bit-sync comparisons and
only widen the genuinely-large-span paths (branch skip, IntCHackCheck, VU
runCycles, the first two VU0 sites).

Two over-widened sites are reverted:
- R5900.cpp cpuTestCycle: (s64) -> (int). The core scheduler predicate;
  making it (s64) while its sibling comparisons stayed (int) caused
  inconsistent event scheduling and the boot hang.
- VU0.cpp _vu0run M-bit break condition: (s64) -> (s32). Only the first two
  VU0 cycle comparisons should be s64; this third one stays s32.

Confirmed on hardware: with the (s64) cast in place, Tekken Tag could not
get past the Namco screen.

+++ pcsx2/R5900.cpp
@@ -192,7 +192,7 @@ __fi int cpuTestCycle(u64 startCycle, s32 delta)
 {
 	// typecast the conditional to signed so that things don't explode
 	// if the startCycle is ahead of our current cpu cycle.
-	return (s64)(cpuRegs.cycle - startCycle) >= delta;
+	return (int)(cpuRegs.cycle - startCycle) >= delta;
 }
 
 static __fi void TESTINT(u8 n, void (*callback)())
